% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regnet-package.R
\docType{package}
\name{regnetMC-package}
\alias{regnetMC}
\alias{regnetMC-package}
\title{regnetMC: Network-Based Regularization for Generalized Linear Models}
\description{
Network-based regularization has achieved success in variable selection for high-dimensional biological data due to its ability to incorporate correlations among genomic features. This package provides procedures of network-based variable selection for generalized linear models (Ren et al. (2017) \doi{10.1186/s12863-017-0495-5} and Ren et al. (2019) \doi{10.1002/gepi.22194}). Two recent additions are the robust network regularization for the survival response and the network regularization for continuous response. Functions for other regularization methods will be included in the forthcoming upgraded versions.
}
\keyword{overview}
